Split e2e out into separate file
