+++ package.json
@@ -6,7 +6,6 @@
   "license": "MIT",
   "scripts": {
     "build": "tsx ./scripts/generate-manifest.ts && webpack --mode production",
-    "build:e2e": "tsx ./scripts/generate-manifest.ts && webpack --mode production",
     "clean": "rm -rf dist/*",
     "lint": "eslint .",
     "lint:ci": "eslint . --quiet",
