+++ playwright/tests/scrapers/cooking.nytimes.com.spec.ts
@@ -33,24 +33,3 @@ testValues.forEach((testValue) => {
     expect(recipe).toEqual(testValue.fixture);
   });
 });
-
-test("open page", async ({ context, extensionId, gotoWithTabIdHelper }) => {
-  const recipeUrl = "https://cooking.nytimes.com/recipes/7131-salmon-burgers";
-
-  const newTabPromise = context.waitForEvent("page");
-
-  const tabId = await gotoWithTabIdHelper(recipeUrl);
-
-  const newTab = await newTabPromise;
-  await newTab.waitForURL(recipeUrl);
-  await newTab.waitForLoadState();
-
-  const extensionPopupPage = await context.newPage();
-  await extensionPopupPage.goto(
-    `chrome-extension://${extensionId}/popup.html?active-tab-id=${tabId}`,
-  );
-
-  await expect(
-    extensionPopupPage.locator('[data-testid="recipe-title"]'),
-  ).toHaveText("Salmon Burgers");
-});
